requirements.txt
neupy==0.6.5
plotly_express==0.4.1
seaborn==0.11.1
numpy==1.19.5
streamlit==0.79.0
streamlit_drawable_canvas==0.6.0
matplotlib==3.3.4
pandas==1.2.3
Pillow==8.2.0
scikit_learn==0.24.2
click<=8.0.4

C:\Users\ME\AppData\Local\Packages\PythonSoftwareFoundation.Python.3.11_qbz5n2kfra8p0\LocalCache\local-packages\Python311\Scripts